more minor fixes
